Fix unicode in title
When downloading a video with a title containing Russian text, the script crashed.  I don't know how to check for the presence of unicode codes in the title and I still need to remove emoticons from the title.  Or use a random name for YouTube-dl so that there are no unicode codes at all